requirements.txt
streamlit==1.39.0
cryptography==43.0.3
pycryptodome==3.20.0